Metal: Change approach to adding ptr arg ASs. (#340)
Turns out that using module cloning isn't working here. If we change
the address space of the arguments, and just remap those arguments
using a value mapper, those changes will get lost after, e.g., a
bitcast from the argument pointer to a different type (that destination
type won't have an address space, resulting in invalid operations).

We were working around this limitation by using a type remapper,
rewriting all pointer types to contain an address space, but that's
obviously invalid: Stack space (i.e. `alloca`s) should still reside
in address space 0.

So instead we're taking the approach we're using with byval lowering,
storing the arguments to a temporary slot. We're still changing the
function's pointer arguments to include an address space, but after
loading the argument and storing it in a stack slot we can dereference
it and get a valid pointer without an address space. We can then
re-use the existing IR with that pointer, without having to rewrite it.

This simplifies the pass a lot, because we don't have to worry about
rewriting intrinsics like memcpy.