Separate codeinfo cache for interpreter params
Currently each codeinfo cache assumes that the same interpreter
parameters are used throughout.  This is problematic when we want to
change the interpreter parameters on a per job basis.

Using a dictionary keyed on the interpreter params to store multiple
codeinfo caches enables changing the interpreter parameters for
different kernel compilations.